Added a new agregate, uses cases and tests

+++ domain/src/main/java/restaurante/order/values/Amount.java
@@ -5,15 +5,15 @@ import co.com.sofka.domain.generic.ValueObject;
 import java.math.BigDecimal;
 import java.util.Objects;
 
-public class Amount implements ValueObject<BigDecimal> {
-    private final BigDecimal value;
+public class Amount implements ValueObject<Integer> {
+    private final Integer value;
 
-    public Amount(BigDecimal value) {
+    public Amount(Integer value) {
         this.value = Objects.requireNonNull(value);
     }
 
     @Override
-    public BigDecimal value() {
+    public Integer value() {
         return this.value;
     }
 }
