+++ domain/src/main/java/restaurante/order/values/Quantity.java
@@ -2,18 +2,17 @@ package restaurante.order.values;
 
 import co.com.sofka.domain.generic.ValueObject;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
-public class Quantity implements ValueObject<BigDecimal> {
-    private final BigDecimal value;
+public class Quantity implements ValueObject<Integer> {
+    private final Integer value;
 
-    public Quantity(BigDecimal value) {
+    public Quantity(Integer value) {
         this.value = Objects.requireNonNull(value);
     }
 
     @Override
-    public BigDecimal value() {
+    public Integer value() {
         return this.value;
     }
 }
